Move compaction strategy code under creation package
