+++ java/compaction/compaction-core/src/main/java/sleeper/compaction/core/job/creation/strategy/CompactionStrategyIndex.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.core.strategy;
+package sleeper.compaction.core.job.creation.strategy;
 
 import sleeper.core.partition.Partition;
 import sleeper.core.statestore.FileReference;
