+++ java/compaction/compaction-core/src/main/java/sleeper/compaction/core/job/creation/strategy/ShouldCreateJobsStrategy.java
@@ -13,19 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.core.strategy;
+package sleeper.compaction.core.job.creation.strategy;
 
-import sleeper.compaction.core.job.CompactionJob;
-import sleeper.compaction.core.job.CompactionJobFactory;
-import sleeper.compaction.core.strategy.CompactionStrategyIndex.FilesInPartition;
+import sleeper.compaction.core.job.creation.strategy.CompactionStrategyIndex.FilesInPartition;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TableProperties;
 
-import java.util.List;
+public interface ShouldCreateJobsStrategy {
 
-public interface LeafPartitionCompactionStrategy {
+    default void init(InstanceProperties instanceProperties, TableProperties tableProperties) {
+    }
 
-    void init(InstanceProperties instanceProperties, TableProperties tableProperties, CompactionJobFactory factory);
+    long maxCompactionJobsToCreate(FilesInPartition filesInPartition);
 
-    List<CompactionJob> createJobsForLeafPartition(FilesInPartition filesInPartition);
+    static ShouldCreateJobsStrategy yes() {
+        return (filesInPartition) -> Long.MAX_VALUE;
+    }
 }
