+++ java/compaction/compaction-core/src/main/java/sleeper/compaction/core/job/creation/strategy/impl/SizeRatioShouldCreateJobsStrategy.java
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.core.strategy.impl;
+package sleeper.compaction.core.job.creation.strategy.impl;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import sleeper.compaction.core.strategy.CompactionStrategyIndex.FilesInPartition;
-import sleeper.compaction.core.strategy.ShouldCreateJobsStrategy;
+import sleeper.compaction.core.job.creation.strategy.CompactionStrategyIndex.FilesInPartition;
+import sleeper.compaction.core.job.creation.strategy.ShouldCreateJobsStrategy;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.properties.table.TableProperty;
