+++ java/compaction/compaction-core/src/test/java/sleeper/compaction/core/strategy/impl/CompactionStrategyTestBase.java
@@ -17,8 +17,8 @@ package sleeper.compaction.core.strategy.impl;
 
 import sleeper.compaction.core.job.CompactionJob;
 import sleeper.compaction.core.job.CompactionJobFactory;
-import sleeper.compaction.core.strategy.CompactionStrategy;
-import sleeper.compaction.core.strategy.CompactionStrategyIndex;
+import sleeper.compaction.core.job.creation.strategy.CompactionStrategy;
+import sleeper.compaction.core.job.creation.strategy.CompactionStrategyIndex;
 import sleeper.core.partition.Partition;
 import sleeper.core.partition.PartitionTree;
 import sleeper.core.partition.PartitionsBuilder;
