+++ java/compaction/compaction-job-creation/src/main/java/sleeper/compaction/job/creation/CreateCompactionJobs.java
@@ -21,9 +21,9 @@ import org.slf4j.LoggerFactory;
 import sleeper.compaction.core.job.CompactionJob;
 import sleeper.compaction.core.job.CompactionJobFactory;
 import sleeper.compaction.core.job.CompactionJobStatusStore;
+import sleeper.compaction.core.job.creation.strategy.CompactionStrategy;
+import sleeper.compaction.core.job.creation.strategy.CompactionStrategyIndex;
 import sleeper.compaction.core.job.dispatch.CompactionJobDispatchRequest;
-import sleeper.compaction.core.strategy.CompactionStrategy;
-import sleeper.compaction.core.strategy.CompactionStrategyIndex;
 import sleeper.compaction.job.creation.commit.AssignJobIdQueueSender;
 import sleeper.compaction.job.creation.commit.AssignJobIdToFiles;
 import sleeper.core.partition.Partition;
