+++ java/compaction/compaction-job-creation/src/test/java/sleeper/compaction/job/creation/CreateCompactionJobsTest.java
@@ -22,9 +22,9 @@ import org.junit.jupiter.api.Test;
 import sleeper.compaction.core.job.CompactionJob;
 import sleeper.compaction.core.job.CompactionJobFactory;
 import sleeper.compaction.core.job.commit.CompactionJobIdAssignmentCommitRequest;
+import sleeper.compaction.core.job.creation.strategy.impl.BasicCompactionStrategy;
+import sleeper.compaction.core.job.creation.strategy.impl.SizeRatioCompactionStrategy;
 import sleeper.compaction.core.job.dispatch.CompactionJobDispatchRequest;
-import sleeper.compaction.core.strategy.impl.BasicCompactionStrategy;
-import sleeper.compaction.core.strategy.impl.SizeRatioCompactionStrategy;
 import sleeper.compaction.core.testutils.InMemoryCompactionJobStatusStore;
 import sleeper.compaction.job.creation.CreateCompactionJobs.BatchJobsWriter;
 import sleeper.compaction.job.creation.CreateCompactionJobs.GenerateBatchId;
