+++ java/system-test/system-test-suite/src/test/java/sleeper/systemtest/suite/CompactionST.java
@@ -22,7 +22,7 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
-import sleeper.compaction.core.strategy.impl.BasicCompactionStrategy;
+import sleeper.compaction.core.job.creation.strategy.impl.BasicCompactionStrategy;
 import sleeper.core.partition.PartitionsBuilder;
 import sleeper.core.properties.validation.CompactionMethod;
 import sleeper.systemtest.dsl.SleeperSystemTest;
